Modify make file
DKAN 1.12 doesn't seem to add a build-dkan.make file. Since this file does not exist, when building the Drupal via drush make, DKAN profile will not get copied. To solve this problem I need to copy DKAN when building via drush make. There are two ways, copying all the files, in the same way it is happenins with the profile dkan japanese, or getting it from github. I am getting from github. I am unsure, though, if this option is preferred in the amiage context